separated spec files for icon list
