+++ cypress/integration/v2/blocks/icon-list/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Icon List Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
