+++ cypress/integration/v2/blocks/icon-list/other-tests.spec.js
@@ -1,0 +1,21 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+	switchDesign,
+	typeContent,
+} from '.'
+
+describe( 'Icon List Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+	switchDesign,
+	typeContent,
+] ) )
