+++ cypress/integration/v2/blocks/icon-list/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Icon List Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
